refactored sweep util stuff into launch land
